// src/webapp/src/api/listen-and-write/model/complexityEnum.ts
/**
 * WriteFluency.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ComplexityEnum = {
    Beginner: 'Beginner',
    Intermediate: 'Intermediate',
    Advanced: 'Advanced'
} as const;
export type ComplexityEnum = typeof ComplexityEnum[keyof typeof ComplexityEnum];