// src/webapp/src/api/listen-and-write/model/exerciseListItemDtoPagedResultDto.ts
/**
 * WriteFluency.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExerciseListItemDto } from './exerciseListItemDto';


export interface ExerciseListItemDtoPagedResultDto { 
    items?: Array<ExerciseListItemDto> | null;
    totalCount?: number;
    pageNumber?: number;
    pageSize?: number;
    readonly totalPages?: number;
    readonly hasPrevious?: boolean;
    readonly hasNext?: boolean;
}