// src/webapp/src/api/listen-and-write/model/subjectEnum.ts
/**
 * WriteFluency.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SubjectEnum = {
    General: 'General',
    Science: 'Science',
    Sports: 'Sports',
    Business: 'Business',
    Health: 'Health',
    Entertainment: 'Entertainment',
    Tech: 'Tech',
    Politics: 'Politics',
    Food: 'Food',
    Travel: 'Travel'
} as const;
export type SubjectEnum = typeof SubjectEnum[keyof typeof SubjectEnum];